Storage/VMDK: Cleanup, get rid of goto... (part 1, build fix)
svn:sync-xref-src-repo-rev: r110626

+++ src/VBox/Storage/VMDK.cpp
@@ -1,4 +1,4 @@
-/* $Id: VMDK.cpp 63800 2016-09-12 13:44:06Z alexander.eichner@oracle.com $ */
+/* $Id: VMDK.cpp 63801 2016-09-12 13:46:54Z alexander.eichner@oracle.com $ */
 /** @file
  * VMDK disk image, core code.
  */
@@ -1067,7 +1067,7 @@ static int vmdkReadGrainDirectory(PVMDKIMAGE pImage, PVMDKEXTENT pExtent)
                         || !pTmpGT2)
                         rc = VERR_NO_MEMORY;
 
-                    uint32_t i = 0;
+                    size_t i = 0;
                     uint32_t *pGDTmp = pExtent->pGD;
                     uint32_t *pRGDTmp = pExtent->pRGD;
